+++ src/main/java/com/github/johnnyjayjay/discord/commandapi/ICommand.java
@@ -1,5 +1,6 @@
 package com.github.johnnyjayjay.discord.commandapi;
 
+import net.dv8tion.jda.core.entities.Guild;
 import net.dv8tion.jda.core.entities.Member;
 import net.dv8tion.jda.core.entities.TextChannel;
 
@@ -27,7 +28,7 @@ public interface ICommand {
      * @return By default, this returns "No info, description or help set" which will be displayed if someone demands help for this command. An Override
      * should return info about the command, such as usage, description, permissions, aliases...
      */
-    default String info() {
+    default String info(Guild guild) {
         return "No info, description or help set for this command";
     }